# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/devices/{device_id}/software:
    get:
      operationId: listDeviceSoftware
      tags: [software]
      security: [{ adminBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: publisher, in: query, schema: { type: string } }
        - { name: architecture, in: query, schema: { $ref: "#/components/schemas/SoftwareArchitecture" } }
        - { name: present, in: query, schema: { type: string, enum: ["true", "false", all], default: "true" } }
        - { name: version, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [name, version, publisher, first_seen, last_seen], default: name } }
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc], default: asc } }
      responses:
        "200": { description: Paginated endpoint software, content: { application/json: { schema: { $ref: "#/components/schemas/DeviceSoftwareList" } } } }
  /v1/devices/{device_id}/software/changes:
    get:
      operationId: listDeviceSoftwareChanges
      tags: [software]
      security: [{ adminBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: change_type, in: query, schema: { $ref: "#/components/schemas/SoftwareChangeType" } }
      responses:
        "200": { description: Paginated endpoint software changes, content: { application/json: { schema: { $ref: "#/components/schemas/SoftwareChangeList" } } } }
  /v1/software:
    get:
      operationId: listFleetSoftware
      tags: [software]
      security: [{ adminBearer: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: publisher, in: query, schema: { type: string } }
        - { name: architecture, in: query, schema: { $ref: "#/components/schemas/SoftwareArchitecture" } }
        - { name: version, in: query, schema: { type: string } }
      responses:
        "200": { description: Paginated fleet software aggregation, content: { application/json: { schema: { $ref: "#/components/schemas/FleetSoftwareList" } } } }
  /v1/software/{software_identity}/devices:
    get:
      operationId: listSoftwareDevices
      tags: [software]
      security: [{ adminBearer: [] }]
      parameters:
        - { name: software_identity, in: path, required: true, schema: { type: string, pattern: "^[a-f0-9]{64}$" } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: version, in: query, schema: { type: string } }
      responses:
        "200": { description: Endpoints with the application, content: { application/json: { schema: { $ref: "#/components/schemas/SoftwareDeviceList" } } } }
...
components:
  parameters:
    DeviceId:
      name: device_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
    adminBearer:
      type: http
      scheme: bearer
    # Interactive console authentication. Deliberately separate from adminBearer,
    # which is a platform-level machine credential and is never issued to a tenant.
    sessionCookie:
      type: apiKey
      in: cookie
      name: nexora_session
  schemas:
...
    SoftwareArchitecture:
      type: string
      enum: [x64, x86, unknown]
    SoftwareChangeType:
      type: string
      enum: [INSTALLED, REMOVED, VERSION_CHANGED]
...
    DeviceSoftwareList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { type: object, additionalProperties: true } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    SoftwareChangeList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { type: object, additionalProperties: true } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    FleetSoftwareList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { type: object, additionalProperties: true } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    SoftwareDeviceList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { type: object, additionalProperties: true } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...